move parallel pool to spack.util.parallel
